Fix main file naming
